Add ability to plot machine timings orderly
We want to modify the machine timing parameters manually using Bokeh for
visualization and have instant feedback of how close this matches the
model data. For this, we added an input/output facility to
ActionFileInfo. Also added the bokeh application which is quite simple
so far but looks incredibly promising.